sc300-portfolio/enterprise-app-registration-demo/demo-app/requirements.txt
Authlib==1.6.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
flash==1.0.3
idna==3.11
iniconfig==2.3.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
